Add hack.sh - This fixes auth engine config?
For some reason the auth engine config that Vault Config Operator is
creating doesn't work... but running the command in ./hack.sh on the
vault-server pod fixes the engine. Not sure what's going on. Committing
so I have this working snippet. At this time `make install` gets you 90%
of the way there, but you need to run hack.sh contents in vault-server
pod after install for Argo to actually sync.